+++ src/main/java/application/Position.java
@@ -1,4 +1,4 @@
-package application.game;
+package application;
 
 import org.immutables.value.Value;
 
@@ -7,6 +7,5 @@ import java.io.Serializable;
 @Value.Immutable
 public interface Position extends Serializable {
     Integer x();
-
     Integer y();
 }
